+++ MailChimp/Models/Forwards.cs
@@ -8,9 +8,9 @@ namespace MailChimpWrapper.Models
 	public class Forwards
 	{
 		[JsonPropertyName("forwards_count")]
-		public long ForwardsCount { get; set; }
+		public int ForwardsCount { get; set; }
 
 		[JsonPropertyName("forwards_opens")]
-		public long ForwardsOpens { get; set; }
+		public int ForwardsOpens { get; set; }
 	}
 }
