+++ MailChimp/Models/Tag.cs
@@ -5,21 +5,21 @@ using System.Text.Json.Serialization;
 
 namespace MailChimpWrapper.Models
 {
-    public class Tag
-    {
-        [JsonPropertyName("id")]
-        public long Id { get; set; }
+	public class Tag
+	{
+		[JsonPropertyName("id")]
+		public int Id { get; set; }
 
-        /// <summary>
-        /// The name of the tag.
-        /// </summary>
-        [JsonPropertyName("name")]
-        public string Name { get; set; }
+		/// <summary>
+		/// The name of the tag.
+		/// </summary>
+		[JsonPropertyName("name")]
+		public string Name { get; set; }
 
-        /// <summary>
-        /// The status for the tag on the member, pass in active to add a tag or inactive to remove it. Possible values: "inactive" or "active".
-        /// </summary>
-        [JsonPropertyName("status")]
-        public string Status { get; set; }
-    }
+		/// <summary>
+		/// The status for the tag on the member, pass in active to add a tag or inactive to remove it. Possible values: "inactive" or "active".
+		/// </summary>
+		[JsonPropertyName("status")]
+		public string Status { get; set; }
+	}
 }
