+++ MailChimp/Models/Timewarp.cs
@@ -8,27 +8,27 @@ namespace MailChimpWrapper.Models
 	public class Timewarp
 	{
 		[JsonPropertyName("gmt_offset")]
-		public long GmtOffset { get; set; }
+		public int GmtOffset { get; set; }
 
 		[JsonPropertyName("opens")]
-		public long Opens { get; set; }
+		public int Opens { get; set; }
 
 		[JsonPropertyName("last_open")]
 		public string LastOpen { get; set; }
 
 		[JsonPropertyName("unique_opens")]
-		public long UniqueOpens { get; set; }
+		public int UniqueOpens { get; set; }
 
 		[JsonPropertyName("clicks")]
-		public long Clicks { get; set; }
+		public int Clicks { get; set; }
 
 		[JsonPropertyName("last_click")]
 		public string LastClick { get; set; }
 
 		[JsonPropertyName("unique_clicks")]
-		public long UniqueClicks { get; set; }
+		public int UniqueClicks { get; set; }
 
 		[JsonPropertyName("bounces")]
-		public long Bounces { get; set; }
+		public int Bounces { get; set; }
 	}
 }
